Antiword

Wraps the antiword utility. Takes a path to an word file and returns text from the document.

Usage

antiword(file = NULL, format = FALSE)

Arguments

file

path or url to your word file

format

format the output text (-f parameter)

Examples

text <- antiword("https://jeroen.github.io/files/UDHR-english.doc")
cat(text)
#> 
#> Universal Declaration of Human Rights
#>                             Plain English Version
#> 
#>  
#> All people everywhere have the same human rights, which no one can take
#> away. This is the basis of freedom, justice and peace in the world.
#> This Declaration affirms the dignity and worth of all people, and the equal
#> rights of women and men. The rights described here are the common standard
#> for all people everywhere.
#> Every person and nation is asked to support the understanding and respect
#> for these rights, and to take steps to make sure that they are recognised
#> and observed everywhere, for all people.
#>  
#> Article 1: You have the same human rights as everyone else in the world,
#> because you are a human being. These rights cannot be taken away from you.
#> Everybody, no matter who they are or where they live, should be treated
#> with dignity.
#>  
#> Article 2: You should not be treated differently, nor have your rights
#> taken away, because of your race, colour, sex, language, religion or
#> political opinions. Your basic rights should be respected no matter what
#> country you are born in or how rich or poor you are.
#>  
#> Article 3: Everyone has the right to life, liberty and security of person.
#>  
#> Article 4:Human beings must not be owned, bought or sold. No one has the
#> right to enslave anyone else. Slavery is a crime.
#>  
#> Article 5: Torture is forbidden at all times and in all circumstances. No
#> one should suffer treatment or punishment that is cruel or makes him or her
#> feel less than human.
#>  
#> Article 6: Everyone has the right to be treated as a person in the eyes of
#> the law.
#>  
#> Article 7: You have the right to be treated by law in the same way as
#> everyone else. You have the same right to be protected by the laws of your
#> country as anyone else.
#>  
#> Article 8: If your rights under the law are violated by someone else, you
#> have the right to see justice done.
#>  
#> Article 9: You may not be arrested or held in a police station without good
#> reason. You may not be kept out of your own country. If you are detained,
#> you have the right to challenge the detention in a court of law.
#>  
#> Article 10: You have the right to a fair and public hearing if you are ever
#> accused of breaking the law, or if you have to go to court for some other
#> reason. The courts must be independent from the government, qualified to
#> understand the law, and free to make their own decisions.
#>  
#> Article 11: If you are accused of a crime, you have the right to be treated
#> as innocent until you are proved guilty, according to the law. You have the
#> right to a fair and public trial where you are allowed to defend yourself.
#> You can not be tried for doing something, which was not a criminal offence
#> in law at the time it was done.
#>  
#> Article 12: No one has the right to intrude in your private life or to
#> interfere with your home or family without good reason. No one has the
#> right to attack your good name without reason. The law should protect you
#> against such interference.
#>  
#> Article 13: You have the right to move about freely within your country.
#> You also have the right to travel to and from your own country, and to
#> leave any country.
#>  
#> Article 14: If you are forced to flee your home because of human rights
#> abuses, you have the right to seek safety in another country. This right
#> does not apply if you have committed a non-political crime or an act that
#> is not in keeping with the UDHR.
#>  
#> Article 15: You have the right to be treated as a citizen of the country
#> you come from. No one can take away your citizenship, or prevent you from
#> changing your country, without good reason.
#>  
#> Article 16: All adults have the right to marry, regardless of their race,
#> country or religion. Both partners have equal rights in the marriage, and
#> their free and full agreement is needed for the marriage to take place. All
#> families are entitled to protection by the state.
#>  
#> Article 17: You have the right to own goods, land and other property, alone
#> or with other people. No one has the right to take your property away
#> without any good reason.
#>  
#> Article 18: You have the right to hold views on any issue you like without
#> fear of punishment or censure. You also have the right to believe in any
#> religion - or none at all. You have the right to change your religion if
#> you wish, and to practice and teach your religion and beliefs.
#>  
#> Article 19: You have the right to tell people your opinion. You should be
#> able to express your views, however unpopular, without fear of punishment.
#> You have the right to communicate your views within your country and to
#> people in other countries.
#>  
#> Article 20: You have the right to peacefully gather together with other
#> people, in public or private. No one should force you to join any group if
#> you do not wish to.
#>  
#> Article 21: You have the right to take part in the government of your own
#> country directly or by being represented. Everyone has the right to equal
#> access to public service in his or her country. Governments represent the
#> will of the people. Therefore free and fair elections should be held on a
#> regular basis.
#>  
#> Article 22: You have the right to have your basic needs met. Everyone is
#> entitled to live in economic, social and cultural conditions that allow
#> them dignity and let them develop as individuals. All countries should do
#> everything they can to make this happen.
#>  
#> Article 23: You have the right to work in fair and safe conditions and to
#> choose your job. You have the right to be paid enough for a decent standard
#> of living, or to receive supplementary benefits. You also have the right to
#> form or join trade unions to protect your interests.
#>  
#> Article 24: You have the right to time off from work. No one may force you
#> to work unreasonable hours, and you have the right to holidays with pay.
#>  
#> Article 25: Everyone has the right to a decent life, including enough food,
#> clothing, housing, medical care and social services. Society should help
#> those unable to work because they are unemployed, sick, disabled or too old
#> to work. Mothers and children are entitled to special care and assistance.
#>  
#> Article 26: Everyone has the right to an education. In the early years of
#> schooling, it should be free of charge and compulsory. Education at a
#> higher lever should be equally available to everyone on the basis of merit.
#> Education should develop the full human being and increase respect for
#> human rights.
#>  
#> Article 27: No one may stop you from participating in the cultural life of
#> your community. You also have the right to share in the benefits scientific
#> discovery may bring, and the right to have any interests from your
#> scientific, literary or artistic work protected.
#>  
#> Article 28: Human beings have the right to live in the kind of world where
#> their rights and freedoms are respected.
#>  
#> Article 29: We all have a responsibility to the people around us, and we
#> can only develop fully as individuals by taking care of each other. All the
#> rights in the UDHR can be limited only by law and then only if necessary to
#> protect other people's rights, meet society's sense of right and wrong,
#> maintain order, and look after the welfare of society as a whole.
#>  
#> Article 30: There is nothing in the UDHR that justifies any person or state
#> doing anything that takes away from the rights which we are all entitled.
#>